Term1 - Project Details
Impact Analysis on Cost of DTH(cable) fees due to TRAIs new regulations. The study is taken to understand if the customers will have to pay more for their DTH(cable) Fees under the new regulations?

It is estimated that the average cost of cable fees was around 250 Rs and the concern among consumers is now that the cost has gone up 2-3 times with this new regulation.

The EDA is done to see what are the offerings available across pay and paid channels and also study a sample of users to understand what is the average cost and if this is really going to be more than what consumers were paying earlier(250 Rs)? 

The analysis is also done to look for measures on how to reduce the cost.